changed styling
