--- src/components/StudentList/StudentList.css
@@ -1,13 +1,0 @@
-.listSection {
-    background-color: '#7f7eff';
-    color: '#e6e6e6';
-  }
-
-.listSection:hover {
-    background-color: '#00d8a4';
-    color: '#e6e6e6';
-  }
-.nested {
-    background-color: '#7f7eff';
-    color: '#e6e6e6';
-  }
